% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{learner}
\alias{learner}
\title{Learner Optimization}
\usage{
learner(
  Y_source,
  Y_target,
  lambda_1,
  lambda_2,
  step_size,
  max_iter = 100,
  threshold = 0.001,
  r,
  max_value = 10
)
}
\arguments{
\item{Y_source}{Matrix containing the source population data.}

\item{Y_target}{Matrix containing the target population data.}

\item{lambda_1}{Regularization parameter lambda1.}

\item{lambda_2}{Regularization parameter lambda2.}

\item{step_size}{Step size for optimization.}

\item{max_iter}{Maximum number of iterations.}

\item{threshold}{Convergence threshold.}

\item{r}{Rank used for the truncated SVD.}

\item{max_value}{Numeric scalar used to specify the maximum value of the objective function allowed before terminating the algorithm. Specifically, the algorithm will terminate if the value of the objective function exceeds \code{max_value}\eqn{\times \epsilon_0}, where \eqn{\epsilon_0} denotes the value of the objective function at the initial point. This is used to prevent unnecessary computation time after the optimization algorithm diverges.}
}
\value{
A list containing the learner estimate, objective values, convergence criterion, and rank.
}
\description{
Learner Optimization
}
